+++ apps/decodex/src/orchestrator/operator_dashboard/app/activity/child-agent/buckets/metrics.js
@@ -1,0 +1,9 @@
+			function childBucketMetricNumber(bucket, field) {
+				const value = Number(bucket?.[field] || 0);
+
+				return Number.isFinite(value) ? Math.max(0, value) : 0;
+			}
+
+			function childBucketWallSeconds(bucket) {
+				return childBucketMetricNumber(bucket, "wall_seconds");
+			}
