+++ apps/decodex/src/orchestrator/operator_dashboard/app/boot/account-names.js
@@ -1,0 +1,46 @@
+			const ACCOUNT_RANDOM_NAMES = [
+				"Alex",
+				"Avery",
+				"Bailey",
+				"Blake",
+				"Casey",
+				"Charlie",
+				"Clara",
+				"Dana",
+				"Drew",
+				"Eden",
+				"Elliot",
+				"Emery",
+				"Evan",
+				"Finley",
+				"Harper",
+				"Hayden",
+				"Iris",
+				"Jamie",
+				"Jordan",
+				"Kai",
+				"Kendall",
+				"Lane",
+				"Liam",
+				"Logan",
+				"Mason",
+				"Maya",
+				"Mia",
+				"Morgan",
+				"Noah",
+				"Nora",
+				"Owen",
+				"Paige",
+				"Parker",
+				"Quinn",
+				"Reese",
+				"Remy",
+				"Riley",
+				"Rowan",
+				"Sage",
+				"Sasha",
+				"Sidney",
+				"Taylor",
+				"Theo",
+				"Val",
+			];
