+++ apps/decodex/src/orchestrator/operator_dashboard/app/boot/nodes.js
@@ -1,0 +1,50 @@
+			const nodes = {
+				projectTitle: document.getElementById("project-title"),
+				flowStepLabels: [...document.querySelectorAll("[data-flow-step]")],
+				flowCounts: {
+					queue: document.getElementById("flow-queue"),
+					run: document.getElementById("flow-run"),
+					review: document.getElementById("flow-review"),
+					land: document.getElementById("flow-land"),
+				},
+				transportHealth: document.getElementById("transport-health"),
+				noticeDock: document.getElementById("notice-dock"),
+				noticeCount: document.getElementById("notice-count"),
+				noticeLabel: document.getElementById("notice-label"),
+				noticeList: document.getElementById("notice-list"),
+				themeButtons: [...document.querySelectorAll("[data-theme-choice]")],
+				workspace: document.getElementById("workspace"),
+				primaryStack: document.getElementById("primary-stack"),
+				accountModeMeta: document.getElementById("account-mode-meta"),
+				panels: {
+					projects: document.getElementById("projects-panel"),
+					accountPool: document.getElementById("account-pool-panel"),
+					currentLanes: document.getElementById("current-lanes-panel"),
+					programs: document.getElementById("programs-panel"),
+					queue: document.getElementById("queue-panel"),
+					recent: document.getElementById("recent-panel"),
+					review: document.getElementById("review-panel"),
+					worktrees: document.getElementById("worktrees-panel"),
+				},
+				sectionMarkers: {
+					control: document.getElementById("section-marker-control"),
+					projects: document.getElementById("section-marker-projects"),
+					execution: document.getElementById("section-marker-execution"),
+					aftercare: document.getElementById("section-marker-aftercare"),
+				},
+				projectOverview: document.getElementById("project-overview"),
+				projectFilterToggle: document.getElementById("project-filter-toggle"),
+				accountPool: document.getElementById("account-pool"),
+				queuedCandidates: document.getElementById("queued-candidates"),
+				queuedMeta: document.getElementById("queued-meta"),
+				currentLanes: document.getElementById("current-lanes"),
+				currentLanesMeta: document.getElementById("current-lanes-meta"),
+				executionPrograms: document.getElementById("execution-programs"),
+				programsMeta: document.getElementById("programs-meta"),
+				recentRuns: document.getElementById("recent-runs"),
+				recentRunsMeta: document.getElementById("recent-runs-meta"),
+				reviewQueue: document.getElementById("review-queue"),
+				reviewLanesMeta: document.getElementById("review-lanes-meta"),
+				worktrees: document.getElementById("worktrees"),
+				worktreesMeta: document.getElementById("worktrees-meta"),
+			};
